services/application-generator/pyproject.toml
[project]
name = "patentforge-application-generator"
version = "1.0.0"
description = "Patent application generation service for PatentForge"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0,<1.0",
    "uvicorn[standard]>=0.34.0,<1.0",
    "langgraph>=0.3.0,<1.0",
    "anthropic>=0.42.0,<1.0",
    "pydantic>=2.0,<3.0",
    "sse-starlette>=2.0,<3.0",
    "python-docx>=1.1.0,<2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.25.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
